refactor: move habitat-dependent config utils to make_habitat_configs.py
