+++ benchmarks/configs/monty_world_experiments.py
@@ -27,7 +27,6 @@ from tbp.monty.frameworks.config_utils.make_dataset_configs import (
     EnvInitArgsMontyWorldStandardScenes,
     EnvironmentDataloaderPerObjectArgs,
     EvalExperimentArgs,
-    PatchViewFinderMontyWorldMountHabitatDatasetArgs,
     PredefinedObjectInitializer,
     RandomRotationObjectInitializer,
     WorldImageDataloaderArgs,
@@ -35,6 +34,9 @@ from tbp.monty.frameworks.config_utils.make_dataset_configs import (
     WorldImageFromStreamDatasetArgs,
     get_env_dataloader_per_object_by_idx,
     get_object_names_by_idx,
+)
+from tbp.monty.frameworks.config_utils.make_habitat_configs import (
+    PatchViewFinderMontyWorldMountHabitatDatasetArgs,
 )
 from tbp.monty.frameworks.environments import embodied_data as ED
 from tbp.monty.frameworks.environments.two_d_data import NUMENTA_OBJECTS
