+++ benchmarks/configs/ycb_experiments.py
@@ -30,15 +30,17 @@ from tbp.monty.frameworks.config_utils.make_dataset_configs import (
     EnvironmentDataloaderPerObjectArgs,
     EvalExperimentArgs,
     ExperimentArgs,
+    PredefinedObjectInitializer,
+    RandomRotationObjectInitializer,
+    get_env_dataloader_per_object_by_idx,
+    get_object_names_by_idx,
+)
+from tbp.monty.frameworks.config_utils.make_habitat_configs import (
     FiveLMMountHabitatDatasetArgs,
     NoisySurfaceViewFinderMountHabitatDatasetArgs,
     PatchViewFinderMountHabitatDatasetArgs,
     PatchViewFinderMultiObjectMountHabitatDatasetArgs,
-    PredefinedObjectInitializer,
-    RandomRotationObjectInitializer,
     SurfaceViewFinderMountHabitatDatasetArgs,
-    get_env_dataloader_per_object_by_idx,
-    get_object_names_by_idx,
 )
 from tbp.monty.frameworks.environments import embodied_data as ED
 from tbp.monty.frameworks.environments.ycb import (
